+++ bin/server.php
@@ -6,37 +6,6 @@ if (PHP_SAPI !== 'cli') {
     exit(E_USER_ERROR);
 }
 
-error_reporting(-1);
-
-if (function_exists('ini_set')) {
-    @ini_set('display_errors', 1);
-
-    $memoryLimit = trim(ini_get('memory_limit'));
-
-    if ($memoryLimit != -1) {
-        $memoryInBytes = function ($value) {
-            $unit = strtolower(substr($value, -1, 1));
-            $value = (int)$value;
-            switch ($unit) {
-                case 'g':
-                    $value *= 1024;
-                case 'm':
-                    $value *= 1024;
-                case 'k':
-                    $value *= 1024;
-            }
-            return $value;
-        };
-
-        // Increase memory_limit if it is lower than 512M
-        if ($memoryInBytes($memoryLimit) < 512 * 1024 * 1024) {
-            @ini_set('memory_limit', '512M');
-        }
-        unset($memoryInBytes);
-    }
-    unset($memoryLimit);
-}
-
 array_shift($argv);
 $arg = function($idx = 1, $default = null) use ($argv) {
     if (is_array($argv)) {
@@ -69,15 +38,57 @@ $opt = function($find, $default = false) use ($argv) {
     return $value;
 };
 
-require __DIR__ . '/../src/bootstrap.php';
+if ($opt('debug')) {
+    error_reporting(-1);
+}
 
-define('TELEPORT_BASE_PATH', rtrim(getcwd(), '/') . '/');
+if (function_exists('ini_set')) {
+    @ini_set('display_errors', 1);
 
-use Teleport\Server;
+    $memoryLimit = trim(ini_get('memory_limit'));
+
+    if ($memoryLimit != -1) {
+        $memoryInBytes = function ($value) {
+            $unit = strtolower(substr($value, -1, 1));
+            $value = (int)$value;
+            switch ($unit) {
+                case 'g':
+                    $value *= 1024;
+                case 'm':
+                    $value *= 1024;
+                case 'k':
+                    $value *= 1024;
+            }
+            return $value;
+        };
+
+        // Increase memory_limit if it is lower than 512M
+        if ($memoryInBytes($memoryLimit) < 512 * 1024 * 1024) {
+            @ini_set('memory_limit', '512M');
+        }
+        unset($memoryInBytes);
+    }
+    unset($memoryLimit);
+}
 
 try {
-    $server = new Server($arg(1, 1337));
-    $server->run();
+    require_once __DIR__ . '/../src/bootstrap.php';
+
+    define('TELEPORT_BASE_PATH', rtrim(getcwd(), '/') . '/');
+
+    $options = array(
+        'debug' => $opt('debug'),
+        'verbose' => $opt('verbose', true)
+    );
+    if (is_readable('config.php')) {
+        $config = include 'config.php';
+        if (is_array($config)) {
+            $options = array_merge($config, $options);
+        }
+    }
+
+    $server = \Teleport\Server::instance($options);
+    $server->run($arg(1, 1337));
 
     printf("server stopped with exit code 0 in %2.4f seconds" . PHP_EOL, microtime(true) - $start);
     exit(0);