+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+  * Refactor tpl parsing to occur before json_decode
   * Add Packages/GC Action
   * Allow tplBase arg to override value for Extract
   * Run APIRequests in sub-process to avoid constant conflicts